API Detector: Check availability of superclass and interfaces
Lint no longer complains about unknown class references, since
Dalvik is careful not to load these until you actually try to
call methods or access fields on them.

However, it looks like superclasses and interfaces are an exception
to this; when the class itself is loaded, it looks like it attempts
to access the interfaces as well. Therefore, flag these in the API
detector.

Change-Id: I9046f4584fd13981f96291944870e357283fc335