Use tools:context in the OverdrawDetector
The OverdrawDetector needs to figure out which activity a given layout
corresponds to. It does this by looking for layout references in all
the activity classes.

However, we now have a tools:context attribute which is added to new
activities by the templates (and which the IDE refactorings is keeping
up to date, and which is editable by the configuration chooser).

This CL makes lint take this attribute into consideration.

Change-Id: Ibacbfbffcc28ce8f7c964a3ac495505aa75aa519