Add ArrayResourceValue and PluralsResourceValue to layoutlib
This changeset adds new classes ArrayResourceValue and
PluralsResourceValue to layoutlib, and makes the res2
parser populate <array>, <string-array>, <integer-array>,
<plurals> etc populate the ResourceValue instances with
the right values. This can be used by layoutlib to render
more than just the fallback value. (For backwards
compatibility, calling just getValue() on these will return
the first value in the collection).

It also fixes the handling of <declare-styleable> items such
that it will immediately also create @attr ResourceItems for
the children within the <declare-styleable> elements.

Change-Id: I11add5d3860b38d3c9e7ab7b4c788eaf4377e4f0