Make ResourceRepository abstract, allow map to be computed dynamically
This changeset moves most of the functionality of ResourceMerger into
a new abstract class AbstractResourceMerger, which does not itself
maintain a map of items; it delegates map creation and handling to
subclasses.

It then implements most of its generic lookup methods by calling
getMap() or getMap(ResourceType) and operating on that.

This allows for example a new MultiResourceRepository in Studio to
combine resources from for example different resource folders, or for
a module repository to combine resources from different module
dependencies.

This CL preserves the ResourceRepository as a concrete class with an
actual map of items, and the data merger and consumer functionality is
moved into that class, since not all AbstractResourceRepository
implementations need it.

Change-Id: If6100149eb74d5ab358bec17b5fe0b7f5311df61