Fix Resource merging for declare-styleable.
Declare styleable can declare attr resources directly
inside them. These should be accessible to merge consumer that
are repository but not to merges that are written on disk
since we want the attr to stay in the declare-styleable
node.

Other attr inside styleable that are references to platform
attr should not be available to any of the merge consumer.

Change-Id: Ie85aa6b4646374b60cbd0bf299415a1a843d3654