Copy template files from sdk.git to tools/base.git
This change copies the template files for AS/ADT from their previous home in sdk/templates
to tools/base/templates.

Subsequent CLs will update the AS packaging scripts and the SDK build scripts
to point to this new location for the templates. A last CL will remove the templates
from their current location in sdk.git.

Change-Id: Ie4d8de2f3e88a86bc0a56350828d1a2b96c07f71