Update internal gradle plugin.
Some weird issue with Gradle 1.7 makes (non runtime) exceptions thrown
by method of a decorated groovy class be encapsulated by a GradleException
which breaks the existing code.

To fix this, this patch moves the working code into a different groovy
class so that the task class only calls this (task classes are meant to
be lightweight anyway).

Change-Id: Id277146971623b33a20c7496a1d08319420f5d08