Add device id's, migrate devices.xml, and add Nexus 7
This changeset deprecates Device#getName(), and adds getId() and
getDisplayName().

Call getId() to get a stable, internal only id for a device, useful to
for example persist which device is chosen in the layout editor for a
given layout. Call getDisplayName() to get the user visible name of a
device. With this change some special cases in the code to generate
menu items from devices can be removed. Plus unit tests for that code.

In addition, this CL moves the vendor device list into sdklib.jar,
just like the generic device list already is, such that it is updated
with the library itself rather than with each SDK release as a
separate file on disk.

Finally, the new Nexus 7 is added to the device list.

Change-Id: I6430b64cd22ad9f9c0c72ccce662835093dc713b