Unify error handling in the resource merging code
This changeset attempts to unify the error handling in the resource
merging code in sdk-common. The purpose of this is to make sure that
error messages follow a predefined format, and more importantly,
always includes a full path to the source file in question (when
possible). This should help the Gradle output parser in Studio more
often be able to link back to the real error location.

It defines a new exception, MergingException, and retrofits the APIs
to throw and expect this exception rather than IOException,
DuplicateDataException and ConsumerException (and the last of these
two are made new subclasses of MergingException). MergingException
tracks the file, line and column numbers of the error, if available.

Code which parses XML now catches SAXParseException (rather than just
SAXException) which allows us to pull out the line and column
information.

Finally, the MergedResourceWriter now pre-validates file names for
file-based resources (e.g. layouts, menus, etc). This is necessary
because when there is an invalid filename (such as an upper case
character in a layout file name), aapt will abort with an error
message, but that error message does not contain a full path to the
error (in fact, it's worse than that: it creates a wrong path,
hardcoding a "res" as the parent name which is not necessarily the
directory containing the bad file in a Gradle project.)

Change-Id: Ibf03173fa4573b92a1eccbd3f4801d493c30c106