Template manifests use full package
This change adds full package data to the manifest generated by
the templates. Previously the generated item was assumed to reside
in the manifest package.

It also expands package names for tools:context references.

In the future we should switch to a conditional model where
we only expand packages when they're not the manifest (app) package.
Change-Id: Ic06fa8d259ed7f410f64ac2c0704beb01e95c812