Add source reference markers in merged manifests
This changeset adds a new property to the manifest merger which allows
the Gradle plugin to get the merger to insert XML comments in the
merged manifest which points back to the original source manifests
from the merge. This is the same mechanism we are already using for
the resource merger, such that in the IDE we can take an error message
pointing to a merged file and change it to point to the real user
source code instead.

This changeset also moves some code from the resource merger into
SdkUtils such that it can be shared between the manifest merger and
the resource merger, and it also makes the source marker code in the
resource merger optional (since we don't want source links in AAR
resources.)

Change-Id: I5c1d730f9b2b1ae102157fd05129e51c80b8a5bb