Don't allow access to security manager fields from sandbox
Code running in the sandbox could look up the secret credential by
using reflection and setAccessible. (We had a unit test for this,
which looked up the field with Class#getField(), and in that case
setAccessible(true) on the field fails; however, turns out if we look
up the field with getDeclaredField() instead, it works.)

This CL turns off the ability to reflectively look up the declared
fields from within the sandbox, and adds a test.

Thanks to Hironori Tokuta for discovering and reporting this.

Change-Id: Id6c4091a4449390575fcc7be2fd50863e66c1218