Fix proguard again.
This time we make sure to set the output during task creation
so that the task actually runs.

However this requires resetting the internal inJarCounts so
that Proguard include all the inJars in the final output,
even for the in jar that are added after the output.

This is a side effect of Proguard supporting multiple
outputs and therefore requiring the inJar/outJar to be
in a certain order that we can't actually do.

Change-Id: Ia535f469d0494ff4c90f4dfbeae6794fc031ac44