Add instrumentation test support in the ADT project importer
This CL adds support for importing instrumentation test projects when
importing an ADT project. Instrumentation tests are imported and
merged into the current project if the ADT instrumentation project is
in a direct sub directory or in a sibling directory relative to the
imported project.

Second, this CL adds a warning to the import summary if the
destination directory you import into contains a risky character in
the path (e.g.  spaces, apostrophes, etc.). Similarly, if the migrated
ProGuard path contains ${user.home} references, the import summary
lists these paths and explains why it's not supported in Gradle.

Third, this CL allows the importer to import projects that are missing
the actual Eclipse project metadata files (.project and .classpath),
provided they still contain the normal structure (manifest, res and
src). This lets us for example import the SDK samples (and a new unit
test imports the ApiDemos project (and its instrumentation test
project) and makes sure it builds.

Finally, various miscellaneous fixes:
- Unit tests for the above new behaviors, as well as scenarios marked
  as TODO to be tested
- The project summary which lists unhandled files now strips out
  directory subtrees that do not contain any non-directory files
- Paths written into build.gradle string literals are now properly
  escaped (e.g. \\ and \')
- Absolute paths in .jar dependencies are now written as project
  relative paths instead
- Mercurial metadata directories are ignored (was already done for
  .git and .svn)

Change-Id: Iae94404f99f42f7db5694a51f28a5dd700b49fab