adb: Use "-P <port>" option only when using a non-default port
As side effect of an earlier CL, adb is always launched with the
"-P <port>" option. This seems to cause errors for a small number
of users. This CL adds this option only if the port to be used is
not the default port.

Fixes https://code.google.com/p/android/issues/detail?id=65700

Change-Id: If02beba43fafc305c78ca1a3585c9df6839fa5a4