Tweak ADT importer to be more module-import friendly
Note that the importer still doesn't allow in-place imports, so
it can't be used to import a module that is first copied verbatim
into the Gradle project directory, but you can point out an ADT
project and import it into a Gradle project that lives elsewhere,
and all the normal import machinery (for resolving dependencies,
replacing .jars with gradle dependencies etc) should kick in.

Import into existing project is similar to importing into
a new project. To import into a new project you do this:

1. GradleImport importer = new GradleImport();
2. // call setters on importer to set preferences
3. importer.importProjects(projects);
4. importer.exportProject(destDir, false);

To import into an existing project, replace step 4 with
4. importer.exportIntoProject(gradleProjectDir, ...);

Also before step 3, call these two setters:
importer.setImportIntoExisting(true);
importer.setPerModuleRepositories(x)
(where x is the same as the corresponding freemarker template
variable perModuleRepositories passed in to the NewModule
template)

Change-Id: Ic0d7b6e02b55951612aebeb856e8d2e3810b54fd