Change device hash used in the AVD Manager.
The previous implementation hashed on the whole state of
the Device object. However only the properties that are
actually serialized in the AVD config.ini matter, so this
implementation hashes only the hardware properties.

It also uses a proper stable hashing method instead of
relying on the Java hashCode() methods (there are some nice
utilities in Guava for this BTW.)

Finally it removes the "broken" state of AVDs when they are
loaded -- if the device configuration has changed, the AVD
config.ini is updated right away to match the new device
properties. The user doesn't need to manually "repair" the
AVD in this case.

Needs tools/swt change: Ib8a19d42036aea977bbccea3baf381adcb781076

Change-Id: Ie61104130d12927794989947d1a663e8b7b4abbc