Add type and declaration resolution to Lint's Java AST
The AST used by lint, Lombok AST, does not contain type information.
That means code which for example sees this code:

    getContext().checkPermission(name)

can't find out which "checkPermission" method this is. That requires
full type resolution.

Lint has had optional type resolution, but this was only implemented
in Studio/IntelliJ.

This changeset adds type resolution for the ECJ based backend, used
when lint is run from Gradle, is run on the command line, and
(shortly) when running in Eclipse ADT.

It also revamps the APIs around parsing (to do type attribution
efficiently, it needs to process all the sources in one go, rather
than one file at a time), and introduces new APIs to pass back
resolved symbols and type descriptors. For example, to pass back the
method declaration for a resolved method call, Lint used to
manufacture a fake AST (a package declaration and a class declaration
node inside it, followed by a method declaration with no body). That
was both cumbersome and inefficient; there is now a ResolvedMethod
class passed back which has simple methods for client to pull
information that is needed, such as checking the containing class'
type name or the method's argument count and types.

Finally, since the API was changing incompatibly anyway, I took the
opportunity to fix some other API issues that had been bothering me:

- The XML and Java parser abstractions are no longer interfaces, they
  are abstract classes, and I renamed the XML parser from a "DOM"
  parser to an "XML" parser

- I cleaned up the methods to generate location handles (and resolving
  types and declarations) such that clients don't have to access them
  via the parser members on the context

- The main client of type resolution in the past was the string format
  detector; I found some bugs in it that I fixed while also cleaning
  up the code to use the new type resolution APIs

- I also hooked up a couple of other lint detectors to using type
  information to more skip warning about APIs if the given class/method
  is not the one the detector was intended for (see SharedPrefsDetector
  and CheckPermissionDetector)

Change-Id: Iac5ee2e4f519a6028b86ec13a87cef82e49b7a83