Implemented implicit default value handling.
Now attributes can have an implicit default value. When merging with
lower priority xml files, trying to reset those attributes will
generate an error similar to the one output when tools:replace is missing.

Added Validator interface to start validating values (wiring to the
xml loading in a subsequent CL).

complete definition of the default values and validators for all the
manifest elements will be done in a subsequent CL for easier review.

Change-Id: Ida4889fed2b13ae627f887059fa11750f6aa7386