Add unused attribute check to the lint API check
This changeset adds a new type of API check issue, unused attribute,
which is flagged for XML attributes that were introduced later than
the minimum SDK version. These are not errors in the way that a call
to a newer method is (which would result in a crash); referencing one
of these attributes will work on older devices in the sense that those
devices will never look up the value. However, we still want to flag
these for the user such that they know that this value is actually not
used on older platforms (in case the value is critical to the
operation of the widget in the case of a layout).

The API level of an attribute is also used to filter out refeference
warnings. For example, if you have the attribute definition
    android:attribute="@android:attr/reference"
We normally flag @android:attr/reference if it was introduced later
than the current minSdkVersion. However, if the attribute itself was
introduced in that same version, or a later version, then there is no
problem since the value reference will never be triggered where it
isn't defined.

Finally, we add a special case to handle the case of using
@android:attr/dividerHorizontal which requires API 11 for the
android:divider attribute (which has minSdkVersion=1) since that
attribute will actually only be read if the showDividers attribute is
also set. This addresses
  67440: API level error for android:attr/dividerHorizontal

Change-Id: Id708f04747a771bbfe73284608f7d00bb9320e27