Fix false positives in the String format detector
This changeset fixes a couple of problems with false positives for the
string format detector.

First, the getString() method in SharedPreferences should not be
treated as a candidate for resource strings; this fixes issue 67597.

Second, when running the detector incrementally in an IDE, we ended up
creating a separate error for each translation in an XML folder,
meaning that when you hovered over the error you could see dozens of
lines of errors.

Third, when running incrementally, any string would be considered a
formatting string, not just strings containing '%' as is the case for
batch mode.

Finally, also fix a couple of infrastructure issues related to the
recently added type support: the .java visitor needs to call
prepareParse() in non-batch mode to ensure type attribution is
performed; the ECJ setup needs to set a couple of additional fields
related to the language compatibility mode (without this you can get
some weird errors related to generics), and finally make the unit test
infrastructure pass a valid compile SDK to the tests such that types
in android.jar can be resolved.

Change-Id: I3a92abe2e3c8deae8dee36ba497afca7902e0260