Move some of the SDK loading to execution phase.
Before each project would do its own SDK parsing, and
would do it in the configuration phase making the SDK
parsing happen even if gradle was called with non-build
tasks (tasks, dependencies, projects, ...).

This patch clean up the SDK loading process and delays
it to the build by putting it in the 'preBuild' task.

This also create a singleton loader that is shared among
all the projects, so that the SDK is loaded only once no
matter how many modules there are.

Finally this also cleans up the AndroidBuilder instantiation.
Long ago, the builder contained the variant information and
so there was one instance per variant. Nowadays, it doesn't
contain any data, and instead receives it through the
build step methods (this is to support incrementality through
the Gradle Task), though it does contain the build target
info but this is shared for all variants.

Therefore this patch remove the map of variant->builder and
use a single builder for the whole module instead.

Change-Id: I0ad47cff963661c275eecbed647a4584cc50c751