Fix predex issue in case of dexing failure.
We always need to do countdown on the latch even
in case of failure or we'll get threads stuck on the latch.

This solves the problem by deleting the pre-dexed file in case
of failure so that other threads that were about the copy the
file can detect the case and do nothing.

Change-Id: I92037c9b6e706943974fed104ea1bc79abc9c208